+++ apps/nocapd/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @nostrwatch/nocapd
 
+## 1.14.0
+
+### Minor Changes
+
+- 3711c6e: fixed hanging job
+
+### Patch Changes
+
+- Updated dependencies [3711c6e]
+  - @nostrwatch/nocap-every-adapter-default@1.7.0
+  - @nostrwatch/nocap@0.9.2
+
+## 1.13.0
+
+### Minor Changes
+
+- a718902: relay urls with usernames were slipping through, migrations and fix sanitizer
+
+### Patch Changes
+
+- Updated dependencies [a718902]
+  - @nostrwatch/nwcache@0.5.0
+
 ## 1.12.2
 
 ### Patch Changes